+++ src/access/mod.rs
@@ -17,7 +17,7 @@ pub fn configure() -> impl FnOnce(&mut ServiceConfig) {
 #[derive(Clone, Debug, Serialize, Deserialize, ToSchema)]
 pub struct CreateAccessData {
     pub name: String,
-    pub create_by: i16,
+    pub create_by: i32,
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize, ToSchema)]
