+++ src/entity/access_entity.rs
@@ -7,8 +7,8 @@ pub struct AccessEntity {
     pub create_time: String,
     pub update_time: String,
     pub name: String,
-    pub create_by: i16, // 创建的用户id
-    pub status: i16,
+    pub create_by: i32, // 创建的用户id
+    pub status: i8,
 }
 
 crud!(AccessEntity {}, "access");
